requirements.txt
flask >= 2.1.3
pyserial >= 3.5
coloredlogs >= 15.0.1
python_dotenv >= 0.20.0
requests >= 2.28.1